"#7 creates initial playbook for descartes"
The only role it has now is common.

Had to refactor a lot of things to get this working, in addition to the
new files:

- modified `group_vars/all.yml` to use ansible user and public key
- added blocks to become root to lots of different tasks that won't work
as the ansible user --- I've only been using these tasks so far running
ansible-pull as root.
- moved "Create ~/tmp directory" task to roles/common/tasks/main.yml
instead of bash_configuration.yml, because that makes sense
- changed calls to `dnf` module in roles/common to `package` to work
with CentOS 7 and Fedora (yum + dnf)
- adjusted indentation on `gitconfig.j2` to make it look better
- modified `hosts.j2` to fix issues I was having with `descartes`
resolving to some other IP --- had to adjust the order; also added a
conditional to install on `home` and `home_server` profiles
- created roles/laptop --- `fuse-exfat` isn't available on the repos I
have configured for `descartes`, and it doesn't really need it, so it
shouldn't be a "common" package. `fuse-exfat` is really needed for
compatability with cross-platform thumb drives, which I really think I'd
only need on my laptops.